311.3. Displaying Validation Messages(Section 23: Handling Forms with Redux Form)
OK so now to show this error message on the screen we can remove the console log.

And underneath our input I'll place a div and it will print out meta dot air like so.

So now we can save this while flip back over.

And we should see our error messages appear.

All right.

So I see you must enter a title and you must enter the description member.

Our validation is going to be ran on the instant this form is rendered on the screen or anytime that

the user interacts with the form.

So in some cases we are probably not going to want to show an error message to the user right away when

they first look at the form.

Usually we only want to show an error message after they try to submit the form.

And there's something wrong with it.

We'll figure out how we do that.

Just a little bit.

But right now we're just going to try to make sure that the form air message shows up correctly.

OK.

So I see you must enter a title here but as soon as I enter some text in that your message goes away

I see the error message here as well.

So I can enter some text in and the error message goes away as well.

So once again every single time your user interacts with the form the validate function gets called

and it's going to validate all the form inputs OK.

So that's pretty much basic validation.

Now we'll take a quick pause right here when we come back the next section we're going to first make

sure that we only show these error messages to the user after they attempt to submit the form.

So again if I just refresh the page and we are essentially telling the user they're doing something

wrong right away.

That's not really helpful for a user.

So we only one show an error message after the user attempts to submit the form.

So quick pause and we'll sort of take care of that in the next video.