312.1. Showing Errors on Touch(Section 23: Handling Forms with Redux Form)
At the end of the last video we got our air message to print up underneath each field.

And I had mentioned that we probably only want to show this error message whenever a user submits the

form.

I was thinking about that a little bit and I realized that maybe there is a better way that we can choose

to show this error message so rather than saying that we're going to show the error message only when

the user clicks on submit.

I think that we should only show the error message after the user clicks into a field enter some text

and then clicks out of it.

So notice how when I select an input I get that little blue boundary right there that indicates that

this field is focused right now.

When I click out the field so I click anywhere down here I no longer get that blue outline.

So that means that the field is no longer focused we can actually watch for that event and only show

the error message after the user has kind of clicked out of the field.

And so I think that would be a little bit better to communicate errors to the user.

So we don't have to wait for them to click on submit to show any possible errors.

So let's figure out how we can do that.

Now before we do one thing to mention really quickly you've probably noticed how any time I click on

this input I see this autocomplete appear the autocomplete stuff is really annoying to me personally

because it's hiding what I have on the screen from you actually sing it very easily.

So if you want to we can optionally turn off auto complete on those text inputs.

We can find the render input function or find our input tag right there and then we can add on auto

complete equals off like so.

Notice how there is a capital C in auto complete.

Now if I saved this I can flip back over it when I type inside if I don't get any autocomplete or dropped

down right and like at all.

OK so let's get back on topic now we want to make sure that we can somehow only show the error message

after the user clicks in and then clicks out.

So to help you understand how we're going to do that we're going to cancel logged that met a property.

One more time.

It's like an ad in a console log of Metta.

And I'm going to save this.

I'll flip back over and remember we get to consulate's right here because we are showing it to different

fields.

I'm going to expand the second console log which is the meta property first specifically the description.

And then inside of here you'll notice that there is a property called Touched.

It's going to collapse this now knows how touched is false.

And then I'm going to click into description I get another console log.

I'm going to inspect that thing I see that touched is false again.

And then as soon as I went back on my console I got a third console log because I de-selected that input.

Now if I look at this object again I see that now touched is true.

All right.

That was a little bit unclear.

So let's do it again.

I'm going to clear that log or refresh the page.

So I'll start off with a touch to false.

Now if I click in and then click out here is the final console right.

You're in touch just now.

True.

So essentially this meta touch property is going to be a boolean that reflects whether or not a user

has kind of entered into an input and then de-selected it.

So in other words if touched is equal that true we probably want to take that as an opportunity to show

our error message to the user if touched is equal to false then we probably don't want to show the error

message at all because the user hasn't really interacted with that input yet.

All right.

Let's give that a shot.