312.2. Showing Errors on Touch(Section 23: Handling Forms with Redux Form)
So I'm going to remove the console log and then to figure out whether or not we should show this error

message.

We are going to end up having to write a decent amount of logic not only to decide if we want to show

the error message but also to make sure that we render out the error message with some appropriate class

names and styling.

So rather than tucking all that logic directly into this div right here I'm going to define a helper

method on our class called this render air and I'm going to pass into that that meta property.

So right above render input I'll define render air.

This is going to be called with the meta object but you and I only care about the air and touch properties

in there.

So we're going to structure out error and touched like so so inside of here we can say if the user has

touched the form and there is an error message then we want to return some error message to show to

the user.

So when we return this we're going to return some nice looking stuff or nice nicely styled elements

here so I'm going to return a div with a class name of UI error message

and that's going to contain a div with the class name of Hetter.

And then finally inside of that Hetter is where we will actually show the air to the user.

So now inside of render input we are calling render air repassing and meta Mehta's couldn't decide whether

or not to show the air.

And then we will either return or not return some amount of GSX.

Now a quick note here.

When we save this and run it inside the browser we're going to very quickly see an error message.

And I bet you can guess what that error message is going to be related to pretty quickly.

So I'm going to save this and flip back over and I see a message your cannot read properly render air

of undefined.

And it looks like it's trying to select the input right here and say the error is on that line of code

right there.

In fact it's actually on line 20 for me.

So the error is coming from calling this render error.

So what's going wrong here.

Well it's our classic friend.

The problem of context inside of re-act classes or javascript classes to be more precise to remember

render input is a function that we pass off to some other component.

So when render input is called It is going to be called with a unknown value of this.

In other words the keyword this side of render input is going to be unknown to us.

It might be undefined it might be Noal it might be some other component.

Who knows.

But it's definitely not the context that we want for our component.