312.3. Showing Errors on Touch(Section 23: Handling Forms with Redux Form)
So to fix this we're going to do the same thing we've done all along.

All we have to do is turn render input into an arrow function.

So I going to say render input equals and then on the other side of the argument list I'll put in my

arrow like so now we can save this and we can flip back over.

We're gonna get that reload as we go.

And so now when I first see my screen right here my form I don't see any error message I can click in.

And I don't see any error message but if I now de-select that input.

I still don't see any major error message.

And so when that happens you might be thinking Well that didn't work at all.

Well this is a real big gotcha here.

If you open up your chrome console and go to the elements tab you can select the element inspector right

here and then just try clicking on the input right there.

So here is our input right here and right underneath it is our div with UI error message.

And inside there is the actual error message we want to show to the user.

So the problem here is not that the error message isn't showing up.

It technically is showing up.

The problem is the fact that semantic UI by default is going to hide error messages.

So if you click on UI error message right here and then you look at the CSSA panel you'll notice that

it has a display of none.

So this is CSSA that is hiding our error message from the user that's all.

So essentially we just have to work with semantic UI a little bit to make sure that the error message

actually shows up.

Let's figure out how we take care of that in the next video.