313.1. Highlighting Errored Fields(Section 23: Handling Forms with Redux Form)
In the last video we worked on our air messaging just a little bit.

Now it probably appears that the message is not showing up.

But like I said this is all because of semantic UI.

So in order to get that erimus to show up we just have to do a little bit of work here it's not.

I mean the worst thing in the world essentially our forum right here inside the render method.

It just needs to have a class name of air.

If that thing has a class name of error than any error messages inside the form are going to be displayed.

But if it doesn't have a class name for air we're not going to see any errors at all.

They're all going to be hidden by default.

So all we really do here is put on a class name of air and then we can save this flip back over.

And now again I'm going to select one of my text inputs and then click out.

And now I successfully see my message OK.

That's pretty much it.

We can also click into description click out and we see the error message.

So now we can go back and each one enter in some text and the errors go away.

Likewise if I refreshed the page entirely I have a fresh form right here and if I click on submit without

touching either field each field is going to show the error message as well.

OK so that's pretty much it for validation.

So we now know how to show an error message to the user.

Now one last thing I want to do in here very quickly just for fun just to make this field look a little

bit better.

The form look a little bit better if we go back to our Rander input right here.

Notice how we have a div with the class name field.

So if we want to we can actually add a nother class name to the thing of air like so if the div with

the field custom field has also the class name air then the entire text input and the label are going

to show up as red as well.

Is that just kind of helps our user figure out what is wrong precisely inside the form.

Of course we don't want to show those as read all the time so we can do a little bit of logic to decide

whether or not we want to show that error message.