313.2. Highlighting Errored Fields(Section 23: Handling Forms with Redux Form)
So essentially I'm going to remove that and I'll say something like I don't know Konst class name above

and then we'll do a little yes 20:15 syntax here.

So I'm going to use a template string with back ticks as opposed to single quotes.

Now I'll say field and then I'll do a little bit of string interpolation here.

So I'll say if Metta air and touched so if both these things are true or truth values then from this

ternary expression that I put right here I'm going to return this string error.

Otherwise I'm going to return a empty string like so.

So then I can reference that variable for my div class name as opposed to just the plain field.

They'll say a class name like so.

All right so I'll save this and we can flip back over.

And now I don't see any error here by default but when I click on it and click out I not only get the

error message but the field is also appearing red as well.

So it's really clear to the user exactly what's going wrong.

OK.

That looks pretty good.

All right I like this.

I think that we're doing great.

So now the last thing we really have to do is start to think about what really happens when we submit

our form.

Yeah we've got the on submit right here to handle submission.

But remember the entire idea behind this form is that we're going to actually create some stream object

or something that represents a user or streaming video to our application.

So let's take a quick pause right here and we come back the next video we're going to start to figure

out exactly what goes on when we submit our form.