317.1. Creating Streams Through Action Creators{yarn add axios redux-thunk --save}(Section 24: REST-Based React Apps)
Now that we have our AP server put together we need to make sure that any time that the user submits

our form from the stream create components we attempt to make Ajax request or a network request over

to our API running on localhost three thousand and one.

So to make a network request we're going to first define an action creator.

We're going to make sure that we wire up that action creator to our component 3D connect helper.

We're going to call the action creator from on Samit right here.

And then the action creator is going to use accedes to make the network request over to our API.

So let's get to it.

Now we got a little bit of set up to do the first thing I'm going to do is change back over to my react

application running inside the terminal.

I'm going to stop this thing and I'm going to install X-posts and redux thunk because we're going to

need both those to write out an asynchronous action creator.