318.1. Creating a Stream with REST Conventions(Section 24: REST-Based React Apps)
In the video we created a new create stream action creator and then imported that along with the connect

function to our stream create component.

So now we need to somehow wire up this action creator to our component.

But as we just saw it's going to be a little bit challenging to hook up both the connect function and

redux form to this component.

So in Honestly in all honesty here we can absolutely do it but this syntax is going to look just a little

bit nasty.

So to do so we could call connect right here.

Place one set of parentheses and then wrap the entire statement all the way from the opening part of

redux form to the very end inside of another set of parentheses.

So there's the opening one and there's the closing one.

This would work out totally fine right here.