318.2. Creating a Stream with REST Conventions(Section 24: REST-Based React Apps)
But very frankly we will use a slightly modified syntax.

So let me show you another way that we could write this out.

I want to remove the export default statement right there and I'm going to instead create a new variable

called form wrapped like so.

So the idea here is that we were making a form wrapped version of our stream create component and then

now at the bottom of the file we could put a another export default statement place or connect function

and then pass in form wrapped like so.

So that's how we can very easily stack up additional kind of nasty looking functions like this without

ending up with some crazy syntax.

So now getting back on track here we do not yet have a map state to prop's function but could pass in

null as the first argument and then as a second argument I'll put in my action creator of the stream.

Now finally inside of my component itself or find the on submit function and each time that the user

Smit's the form I'm going to make sure that I call this stop prop's dot stream or I forgot the name

of it.

Great stream there we go.

Reate stream and we will call it with all of our foreign values.

It's now whenever the user tries to estimate the form we're going to validate the inputs if the inputs

are valid we will call on submit on Sammet is going to call our action creator of create stream that's

going to run the Create stream action creator right here and we're going to attempt to make a request

over to our API server and create a new stream and we know that this is going to create a stream because

we are following restful conventions.

We are making a post request to slash streams.

All right.

So going to save all this and then recall that we had stopped our server to do a new install.

Back over here and you know what I'm actually looking at the new install We forgot one little stepper.

I forgot to step here.

We forgot to wire up redux thunk.

Let's do that really quickly.

While we do that let's start our server up as well with NPM start OK so well that goes I'm going to

flip back over to my roots index such as file.

So here's the root index file.

Remember inside here we had already set up the applied middle where we call in anticipation of wiring

up redux thunked So this is going to be really easy at the top.

I'm going to import redux thunk from redux dasht thunk and then I will make sure that I pass that into

the applied middleware call right here redux.

Thank all.

All right let's flip back over to our application.

Here we are.

I see my form on the screen so now I'm going to pull up in my network request log.

I'm going to filter by S.H. H.R. requests and I'll enter in some stream names so how about like my stream

and I'll give it a description or something like this is a great stream and then I'll submit.

All right so we see a little nasty error message here.

But I bet you can guess why we see this one.

Yeah.

Hold traditional.