318.3. Creating a Stream with REST Conventions
We've got a callback function that we are passing off to a component and we did not bind it.

So let's make sure that we defined this with an arrow function.

Now I'll be honest with you I did not plan this error.

I plan on the one around render input.

When we ran into it.

But this is just an example of how often you tend to run into this stuff unfortunately.

All right let's save this flip back over again.

I'll say my stream.

This is a great stream.

And then I'll submit it.

And now I see two requests appear right here.

I see an options request which is totally fine.

And then I see a post request that just you know the options request right there is because we are making

requests to a server that is running on a different port.

This is related to corps request errors.

If you are a court enforcement policy if you've ever taken a look at core's errors before so it looks

like Course is working successfully here and we've got our post request that is returning a status of

201 which indicates that our stream was successfully created.

So now we can check to see if that stream was actually made or not to do so.

I can go back over in a new terminal window to the API directory that we had just created with that

DBI just on file.

Now if you are on a Mac OS or a Linux based system you can run cat D-B Jaison and that will print out

the contents of that file and now you'll see that the API server has automatically added a new record

to that file under the streams key.

This is our stream and it is automatically assigned an idea of one.

Now if you are running on Windows and cannot use the cat command you can always start up your code editor

inside this folder and inspect that file and just verify that the stream was created successfully.

OK.

That's pretty much it.

Again because we have followed restful conventions we didn't have to look up complicated documentation

or anything like that.

We just made a post request to the appropriate end point and our record was created for us.

So this is fantastic.

Now that we have the ability to create new streams it's time for us to focus back on the original streamlets

component and make sure that we can show a list of all the streams that have been created over here

as well.

But before we do let's also make sure that we start to put a little bit of thought into wrapping up

our create stream action carried over here.

So we got a couple of things that we're juggling simultaneously.

Let's take a quick pause here and we'll continue in the next video.